Add functionality to Arguments to cross-declare other argument dependencies.
I.E. If an argument X is only useful when combined with Y and Z,  you could define X as `mustBeUsedWith(Set.of("Y", "Z"))` to ensure Y and Z must also be present when X is used, even if Y or Z could be used independently.
Similarly, if you have mutually EXCLUSIVE arguments, you can declare this with `mustNotBeUsedWith(Set.of("X"))` on one of the exclusive args.